+++ core/domain/src/main/java/univ/earthbreaker/namu/core/support/tx/TransactionAction.java
@@ -1,4 +1,4 @@
-package univ.earthbreaker.namu.core.support;
+package univ.earthbreaker.namu.core.support.tx;
 
 @FunctionalInterface
 public interface TransactionAction<T> {
