+++ core/domain/src/main/java/univ/earthbreaker/namu/core/support/tx/TransactionHandler.java
@@ -1,4 +1,4 @@
-package univ.earthbreaker.namu.core.support;
+package univ.earthbreaker.namu.core.support.tx;
 
 import org.springframework.transaction.TransactionException;
