+++ core/domain/src/main/java/univ/earthbreaker/namu/core/support/tx/TransactionHandlerAdapter.java
@@ -1,4 +1,4 @@
-package univ.earthbreaker.namu.core.support;
+package univ.earthbreaker.namu.core.support.tx;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.TransactionException;
